last

+++ src/test/Dirtybytes.sol
@@ -2,6 +2,7 @@
 pragma solidity ^0.8.0;
 
 import "forge-std/Test.sol";
+
 /*
     "description": "Copying ``bytes`` arrays from memory or calldata to storage is done in chunks of 32 bytes even if the length is not a multiple of 32. 
     Thereby, extra bytes past the end of the array may be copied from calldata or memory to storage. 
@@ -11,24 +12,27 @@ import "forge-std/Test.sol";
     
     "link": https://blog.soliditylang.org/2022/06/15/dirty-bytes-array-to-storage-bug/
     "fixed": 0.8.15
-
 */
 
 contract ContractTest is Test {
-        Dirtybytes Dirtybytesontract;
+    Dirtybytes Dirtybytesontract;
 
-function testDirtybytes() public {
- 
-    Dirtybytesontract = new Dirtybytes();
-    emit log_named_bytes("Array element in h() not being zero::",Dirtybytesontract.h());
-    console.log("Such that the byte after the 63 bytes allocated below will be 0x02.");
-
+    function testDirtybytes() public {
+        Dirtybytesontract = new Dirtybytes();
+        emit log_named_bytes(
+            "Array element in h() not being zero::",
+            Dirtybytesontract.h()
+        );
+        console.log(
+            "Such that the byte after the 63 bytes allocated below will be 0x02."
+        );
     }
- 
 }
+
 contract Dirtybytes {
-    event ev(uint[], uint);
+    event ev(uint256[], uint256);
     bytes s;
+
     constructor() {
         // The following event emission involves writing to temporary memory at the current location
         // of the free memory pointer. Several other operations (e.g. certain keccak256 calls) will
@@ -36,12 +40,13 @@ contract Dirtybytes {
         // In this particular case, the length of the passed array will be written to temporary memory
         // exactly such that the byte after the 63 bytes allocated below will be 0x02. This dirty byte
         // will then be written to storage during the assignment and become visible with the push in ``h``.
-        emit ev(new uint[](2), 0);
+        emit ev(new uint256[](2), 0);
         bytes memory m = new bytes(63);
         s = m;
     }
+
     function h() external returns (bytes memory) {
         s.push();
         return s;
-      }
+    }
 }
